requirements.txt
Django==1.6.1
South==0.8.4
Unipath==1.0
argparse==1.2.1
django-braces==1.3.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.0.1
django-filter==0.7
django-model-utils==2.0
django-registration==1.0
django-taggit==0.11.2
django-vanilla-views==1.0.2
django-widget-tweaks==1.3
six==1.5.2
sqlparse==0.1.10
wsgiref==0.1.2
